requirements.txt
appnope==0.1.4
asgiref==3.10.0
asttokens==3.0.0
av==14.4.0
Box2D==2.3.10
certifi==2025.4.26
charset-normalizer==3.4.2
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.5.1
gymnasium==1.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
npm==0.1.1
numpy==2.2.4
opencv-python==4.11.0.86
optional-django==0.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
react==4.3.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
swig==4.3.1
sympy==1.14.0
tailwind==3.1.5b0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
